call to srand only once in the beginning of main, in order for number generator to produce random numbers correctly; INTS changes to FLOATS for distances, and fitness

+++ tsp.h
@@ -3,9 +3,9 @@
 
 using std::vector;
 struct tour{
-int fitness;
+float fitness;
 int* tour;
-int *tour_lengths;
+float *tour_lengths;
 };
 
 struct city
